app_rpt.c: change timers of type long to int The changed timers are unnecessarily of type long.

+++ apps/app_rpt/app_rpt.h
@@ -394,12 +394,12 @@ struct rpt_link {
 	char	disced;
 	char	killme;
 	long	elaptime;
-	long	disctime;
-	long 	retrytimer;
-	long	retxtimer;
-	long	rerxtimer;
-	long	rxlingertimer;
-	int     rssi;
+	int	disctime;
+	int	retrytimer;
+	int	retxtimer;
+	int	rerxtimer;
+	int	rxlingertimer;
+	int	rssi;
 	int	retries;
 	int	max_retries;
 	int	reconnects;
@@ -408,7 +408,7 @@ struct rpt_link {
 	struct ast_channel *pchan;	
 	char	linklist[MAXLINKLIST];
 	time_t	linklistreceived;
-	long	linklisttimer;
+	int	linklisttimer;
 	int	dtmfed;
 	int linkunkeytocttimer;
 	struct timeval lastlinktv;
@@ -778,8 +778,8 @@ struct rpt {
 	int dtmfidx,rem_dtmfidx;
 	int dailytxtime,dailykerchunks,totalkerchunks,dailykeyups,totalkeyups,timeouts;
 	int totalexecdcommands, dailyexecdcommands;
-	long	retxtimer;
-	long	rerxtimer;
+	int	retxtimer;
+	int	rerxtimer;
 	long long totaltxtime;
 	char mydtmf;
 	char exten[AST_MAX_EXTENSION];
@@ -816,7 +816,7 @@ struct rpt {
 	int tailmessagen;
 	time_t disgorgetime;
 	time_t lastthreadrestarttime;
-	long	macrotimer;
+	int	macrotimer;
 	char	lastnodewhichkeyedusup[MAXNODESTR];
 	int	dtmf_local_timer;
 	char	dtmf_local_str[100];
@@ -843,7 +843,7 @@ struct rpt {
 	char newkey;
 	char iaxkey;
 	char inpadtest;
-	long rxlingertimer;
+	int rxlingertimer;
 	char localoverride;
 	char ready;
 	char lastrxburst;
